make storage_size  being u64

+++ srml/contract/src/account_db.rs
@@ -120,7 +120,7 @@ impl<T: Trait> AccountDb<T> for DirectAccountDb {
 				} else if let Some(code_hash) = changed.code_hash {
 					AliveContractInfo::<T> {
 						code_hash,
-						storage_size: <Module<T>>::storage_size_offset(),
+						storage_size: <Module<T>>::storage_size_offset() as u64,
 						trie_id: <T as Trait>::TrieIdGenerator::trie_id(&address),
 						deduct_block: <system::Module<T>>::block_number(),
 						rent_allowance: <BalanceOf<T>>::max_value(),
@@ -145,10 +145,10 @@ impl<T: Trait> AccountDb<T> for DirectAccountDb {
 
 				for (k, v) in changed.storage.into_iter() {
 					if let Some(value) = child::get_raw(&new_info.trie_id[..], &blake2_256(&k)) {
-						new_info.storage_size -= value.len() as u32;
+						new_info.storage_size -= value.len() as u64;
 					}
 					if let Some(value) = v {
-						new_info.storage_size += value.len() as u32;
+						new_info.storage_size += value.len() as u64;
 						child::put_raw(&new_info.trie_id[..], &blake2_256(&k), &value[..]);
 					} else {
 						child::kill(&new_info.trie_id[..], &blake2_256(&k));
